atvec: Added --thresh and --attribute-thresholds
atvec can now compute thresholds (--thresh) which give the best accuracy
on the training set. This is based on a binary search across
possible values. This assumes monotonic behavior, which isn't
always the case - but in testing seems to produce good results.

These thresholds can then be used in --roc with the optional
--attribute-thresholds option (default is still zero). The
histograms have been modified to show graphically the threshold
in the graph.